EP 24 Authorization With Policies

+++ app/Http/Controllers/ThreadController.php
@@ -162,6 +162,17 @@ class ThreadController extends Controller
         // Option 1 see create_replies_table
         /* $thread->replies()->delete(); */
 
+        $this->authorize('update', $thread);
+        
+        // if ($thread->user_id != auth()->id()) {
+        //     abort(403, 'You do not have permission to do this.');
+            /* if (request()->wantsJson()) {
+                return response(['status' => 'Permission Denied'], 403);
+            }
+
+            return redirect('/login'); */
+        // }
+
         // Option 3rd, overwrite the delete() method.
         $thread->delete();
